Update Gitlab CI
Add Go v1.16 to the test matrix.

Replace the separate download and testing of lint tools in the test job
with golangci-lint.  The exception is misspell, which is also used to
check markdown files.

Add clang 10 to the test matrix

Refactored jobs to combine all clang jobs running on debian buster.  Now
using a parallel matrix.  Similar for clang jobs running on stretch.