Fix 2 bugs in the HTML reports
The javascript for sorting tables was tied to the exact columns present
in the table, but the columns included in the table depend on what
coverage data is present.  The updated javascript does not depend on
the columns knowing theiri position.

Also fixed a small style bug, where the section "By Function" was
present, even if function coverage data was not present.